Fix: Support decimal scores in evaluation forms and PDF
Changes:
- Changed all score columns from INTEGER to FLOAT in models.py
- Updated PDF generator to format decimal scores (3.5, 7.5) correctly
- Added migration script to convert existing database
- Frontend already had step="0.5" for decimal input support

This allows evaluators to give half-point scores like:
- 3.5 / 10
- 7.5 / 15
- 82.5 / 100 (total)

+++ backend/models.py
@@ -165,72 +165,72 @@ class FicheEvaluation(db.Model):
     # II - RESULTATS DE L'EVALUATION - Critères avec barèmes exacts du formulaire réel
     
     # PERTINENCE (/5)
-    pertinence_score = db.Column(db.Integer, default=0)  # 0-5
+    pertinence_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     pertinence_description = db.Column(db.Text)
     pertinence_recommandations = db.Column(db.Text)
 
     # ALIGNEMENT A LA DOCTRINE DE TRANSFORMATION SYSTEMIQUE (/10)
-    alignement_score = db.Column(db.Integer, default=0)  # 0-10
+    alignement_score = db.Column(db.Float, default=0)  # 0-10 (permet les décimales)
     alignement_description = db.Column(db.Text)
     alignement_recommandations = db.Column(db.Text)
 
     # PERTINENCE DES ACTIVITES ET BIEN FONDE DES COUTS/PART DE FONCTIONNEMENT (/15)
-    activites_couts_score = db.Column(db.Integer, default=0)  # 0-15
+    activites_couts_score = db.Column(db.Float, default=0)  # 0-15 (permet les décimales)
     activites_couts_description = db.Column(db.Text)
     activites_couts_recommandations = db.Column(db.Text)
 
     # ÉQUITE (SOCIALE-TERRITORIALE-GENRE) (/15)
-    equite_score = db.Column(db.Integer, default=0)  # 0-15
+    equite_score = db.Column(db.Float, default=0)  # 0-15 (permet les décimales)
     equite_description = db.Column(db.Text)
     equite_recommandations = db.Column(db.Text)
 
     # VIABILITE/RENTABILITE FINANCIERE (/5)
-    viabilite_score = db.Column(db.Integer, default=0)  # 0-5
+    viabilite_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     viabilite_description = db.Column(db.Text)
     viabilite_recommandations = db.Column(db.Text)
 
     # RENTABILITE SOCIO-ECONOMIQUE (ACA/MPR) (/5)
-    rentabilite_score = db.Column(db.Integer, default=0)  # 0-5
+    rentabilite_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     rentabilite_description = db.Column(db.Text)
     rentabilite_recommandations = db.Column(db.Text)
 
     # BENEFICES STRATEGIQUES (SECURITE-RESILIENCE-INNOVATION-COMPETITIVITE-CONTENU LOCAL, ETC.) (/15)
-    benefices_strategiques_score = db.Column(db.Integer, default=0)  # 0-15
+    benefices_strategiques_score = db.Column(db.Float, default=0)  # 0-15 (permet les décimales)
     benefices_strategiques_description = db.Column(db.Text)
     benefices_strategiques_recommandations = db.Column(db.Text)
 
     # PERENNITE ET DURABILITE DES EFFETS ET IMPACTS DU PROJET (/5)
-    perennite_score = db.Column(db.Integer, default=0)  # 0-5
+    perennite_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     perennite_description = db.Column(db.Text)
     perennite_recommandations = db.Column(db.Text)
 
     # AVANTAGES ET COUTS INTANGIBLES (/10)
-    avantages_intangibles_score = db.Column(db.Integer, default=0)  # 0-10
+    avantages_intangibles_score = db.Column(db.Float, default=0)  # 0-10 (permet les décimales)
     avantages_intangibles_description = db.Column(db.Text)
     avantages_intangibles_recommandations = db.Column(db.Text)
 
     # FAISABILITE DU PROJET / RISQUES POTENTIELS (/5)
-    faisabilite_score = db.Column(db.Integer, default=0)  # 0-5
+    faisabilite_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     faisabilite_description = db.Column(db.Text)
     faisabilite_recommandations = db.Column(db.Text)
 
     # POTENTIALITE OU OPPORTUNITE DU PROJET A ETRE REALISE EN PPP (/5)
-    ppp_score = db.Column(db.Integer, default=0)  # 0-5
+    ppp_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     ppp_description = db.Column(db.Text)
     ppp_recommandations = db.Column(db.Text)
 
     # IMPACTS ENVIRONNEMENTAUX (/5)
-    impact_environnemental_score = db.Column(db.Integer, default=0)  # 0-5
+    impact_environnemental_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     impact_environnemental_description = db.Column(db.Text)
     impact_environnemental_recommandations = db.Column(db.Text)
-    
+
     # IMPACT SUR L'EMPLOI (/5) - Ajouté selon demande utilisateur
-    impact_emploi_score = db.Column(db.Integer, default=0)  # 0-5
+    impact_emploi_score = db.Column(db.Float, default=0)  # 0-5 (permet les décimales)
     impact_emploi_description = db.Column(db.Text)
     impact_emploi_recommandations = db.Column(db.Text)  # Recommandations spécifiques pour l'emploi
-    
+
     # Score total sur 105 (avec l'ajout de l'impact emploi)
-    score_total = db.Column(db.Integer, default=0)  # Somme de tous les scores
+    score_total = db.Column(db.Float, default=0)  # Somme de tous les scores (permet les décimales)
     
     # III - CONCLUSION
     proposition = db.Column(db.String(50))  # "Favorable", "Favorable sous condition", "Défavorable", "Compléments demandés"
